@@ -11,7 +11,7 @@
   },
   "license": "ISC",
   "dependencies": {
-    "next": "15.0.0-rc.1",
+    "next": "15.3.1",
     "next-auth": "workspace:*",
     "react": "19.0.0-rc-4c58fce7-20240904",
     "react-dom": "19.0.0-rc-4c58fce7-20240904"
