@@ -1,15 +1,15 @@
 // @ts-check
 
-const { MarkdownPageEvent } = require("typedoc-plugin-markdown")
-const path = require("path")
-const fs = require("fs")
+import { MarkdownPageEvent } from "typedoc-plugin-markdown"
+import path from "path"
+import fs from "fs"
 
 /**
  * Local plugin to tweak TypeDoc output for nextra docs
  *
  *  @param {import("typedoc-plugin-markdown").MarkdownApplication} app
  */
-module.exports.load = (app) => {
+export function load(app) {
   injectNextraCalloutImport(app)
   parseOutput(app)
   writeMetaJsFiles(app)
@@ -57,7 +57,7 @@ function parseOutput(app) {
     const replaceCodeBlockTitle = (match, p1, p2, p3) => `${p1}filename="${p3}"`
 
     page.contents = page.contents
-      .replace(calloutRegex, replaceCallout)
+      ?.replace(calloutRegex, replaceCallout)
       .replace(codeBlockRegex, replaceCodeBlockTitle)
   })
 }
@@ -82,7 +82,7 @@ function writeMetaJsFiles(app) {
     ) => {
       const pages = defaultValue
       navigationItems.forEach((item) => {
-        const pageKey = item.url ? path.parse(item.url).name : null
+        const pageKey = item.path ? path.parse(item.path).name : null
         if (pageKey) {
           pages[pageKey] = item.title
           if (item?.children && item?.children?.length > 0) {
@@ -101,15 +101,19 @@ function writeMetaJsFiles(app) {
 
       const metaJString = `
 export default ${JSON.stringify(pages, null, 2)}`
+
       if (new RegExp(".*docs/pages/reference$").test(outputDirectory)) return
+
       fs.writeFileSync(path.join(outputDirectory, "_meta.js"), metaJString)
     }
 
     /**
      * Recursively write _meta.js files for each page based on output.navigation
      */
-    writeMetaJs(output.navigation, output.outputDirectory, {
-      overview: "Overview",
-    })
+    if (output.navigation) {
+      writeMetaJs(output.navigation, output.outputDirectory, {
+        overview: "Overview",
+      })
+    }
   })
 }
