@@ -10,11 +10,6 @@ export default {
   express: "@auth/express",
   qwik: "@auth/qwik",
   "solid-start": "@auth/solid-start",
-  warnings: "Warnings",
-  errors: {
-    title: "Errors",
-    href: "/reference/core/errors",
-  },
   "--- adapters": {
     type: "separator",
     title: "Adapters",
@@ -33,6 +28,7 @@ export default {
   "mikro-orm-adapter": "@auth/mikro-orm-adapter",
   "mongodb-adapter": "@auth/mongodb-adapter",
   "neo4j-adapter": "@auth/neo4j-adapter",
+  "neon-adapter": "@auth/neon-adapter",
   "pg-adapter": "@auth/pg-adapter",
   "pouchdb-adapter": "@auth/pouchdb-adapter",
   "sequelize-adapter": "@auth/sequelize-adapter",
