@@ -24,15 +24,16 @@ module.exports = {
   tsconfig: "./tsconfig.json",
   plugin: [
     "typedoc-plugin-markdown",
-    require.resolve("./typedoc-nextauth.cjs"),
+    require.resolve("./typedoc-nextauth.js"),
     "typedoc-plugin-mdn-links",
+    "typedoc-plugin-no-inherit",
   ],
   disableSources: true,
   excludeNotDocumented: true,
   excludeExternals: true,
   excludeInternal: true,
-  excludePrivate: true,
   excludeProtected: true,
+  excludeReferences: true,
   cleanOutputDir: false,
   gitRevision: "main",
   githubPages: false,
@@ -75,8 +76,5 @@ module.exports = {
   expandObjects: true,
   parametersFormat: "table",
   indexFormat: "table",
-  textContentMappings: {
-    "label.packages": "Integrations",
-  },
   useCodeBlocks: true,
 }
